[Sketcher][planegcs] Handle type issues with `int` and `size_t`
... and `unsigned int`

Suggestions thanks to @abdullahtahiriyo

+++ src/Mod/Sketcher/App/planegcs/GCS.h
@@ -257,7 +257,7 @@ namespace GCS
                                         int tagId=0, bool driving = true);
         int addConstraintTangentCircumf(Point &p1, Point &p2, double *rd1, double *rd2,
                                         bool internal=false, int tagId=0, bool driving = true);
-        int addConstraintTangentAtBSplineKnot(BSpline &b, Line &l, size_t knotindex,
+        int addConstraintTangentAtBSplineKnot(BSpline &b, Line &l, unsigned int knotindex,
                                               int tagId=0, bool driving = true);
 
         // derived constraints
